chore: Move switch label generation to own file 🧹
